Move request_refund_payment to flow.rs
Move request_refund_payment documentation to flow.rs
